Safe another actions so not authorize agency admin cant change other agencies data

+++ Controllers/GalleryController.cs
@@ -1,4 +1,5 @@
 using Book.App.Filters.Exception;
+using Book.App.Helpers;
 using Book.App.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -10,17 +11,20 @@ public class GalleryController : Controller
 {
     private readonly ITourService _tourService;
     private readonly IImageService _imageService;
+    private readonly IAgencyService _agencyService;
 
-    public GalleryController(ITourService tourService, IImageService imageService)
+    public GalleryController(ITourService tourService, IImageService imageService, IAgencyService agencyService)
     {
         _tourService = tourService;
         _imageService = imageService;
+        _agencyService = agencyService;
     }
 
     [Authorize(Roles = "Admin,AgencyAdmin")]
     // GET: Gallery{tourId}
-    public async Task<IActionResult> Edit(int id)
+    public async Task<IActionResult> Edit(int id, int agencyId)
     {
+        if (!await _agencyService.IsUserInAgencyAsync(this.GetCurrentUserId(), agencyId)) throw new NotInAgencyException();
         var tour = await _tourService.GetById(id);
         return View(tour);
     }
@@ -29,12 +33,12 @@ public class GalleryController : Controller
     // POST: Gallery{tourId}
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Add(List<IFormFile> files, int id)
+    public async Task<IActionResult> Add(List<IFormFile> files, int id, int agencyId)
     {
+        if (!await _agencyService.IsUserInAgencyAsync(this.GetCurrentUserId(), agencyId)) throw new NotInAgencyException();
         await _imageService.AddImagesToTour(files, id);
-        return RedirectToAction("Edit", new { id });
+        return RedirectToAction("Edit", new { id, agencyId });
     }
-
 
     [Authorize]
     // GET Details
@@ -56,9 +60,10 @@ public class GalleryController : Controller
 
     [Authorize(Roles = "Admin,AgencyAdmin")]
     // POST: Gallery/Delete{imageId}
-    public async Task<IActionResult> Delete(int id, int tourId)
+    public async Task<IActionResult> Delete(int id, int tourId, int agencyId)
     {
+        if (!await _agencyService.IsUserInAgencyAsync(this.GetCurrentUserId(), agencyId)) throw new NotInAgencyException();
         await _imageService.Delete(id);
-        return RedirectToAction("Edit", new { id = tourId });
+        return RedirectToAction("Edit", new { id = tourId, agencyId });
     }
 }
